Add patch to not touch state while powering on
This patch prevents nhc from changing node state in Slurm while nodes
are powering on and while NHC is running before slurmd is registering
to slurmctld. Old versions of slurm were reporting node as idle~ in
this case, latest version of slurm report mix~ or alloc~.

This patch is designed to support these new statuses.