Revert "Change default installation paths"
This reverts commit b75edbb822c2173a4f5589f7ea366b97ef703854. This
commit breaks the build of the RPM package, as the macro
nhc_helper_dir must be changed according to the libexecdir for the
build to succeed.

Also, the package currently in production on EDF clusters does not use
these modified paths, it uses the default paths. These modified have
never been used, then revert to the default.